add repository for Data, and some ingections

+++ BakingApp/app/src/main/java/com/bakingapp/android/data/Ingredient.java
@@ -11,7 +11,7 @@ public class Ingredient implements Parcelable{
 
     @SerializedName("quantity")
     @Expose
-    private Integer quantity;
+    private Float quantity;
     @SerializedName("measure")
     @Expose
     private String measure;
@@ -19,11 +19,11 @@ public class Ingredient implements Parcelable{
     @Expose
     private String ingredient;
 
-    public Integer getQuantity() {
+    public Float getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(Float quantity) {
         this.quantity = quantity;
     }
 
@@ -59,7 +59,7 @@ public class Ingredient implements Parcelable{
     }
 
     protected Ingredient(Parcel in) {
-        this.quantity = (Integer) in.readValue(Integer.class.getClassLoader());
+        this.quantity = (Float) in.readValue(Float.class.getClassLoader());
         this.measure = in.readString();
         this.ingredient = in.readString();
     }
